Refactoring da estrutura do projeto como um todo. Sensornode não compila.
