--- bodyhub/CMakeLists.txt
@@ -1,55 +1,0 @@
-CMAKE_MINIMUM_REQUIRED (VERSION 2.8)
-
-PROJECT (bodyhub)
-
-###########################################################################
-# Set the search path for .cmake files.
-SET (CMAKE_MODULE_PATH "${CMAKE_CURRENT_SOURCE_DIR}/cmake.Modules" ${CMAKE_MODULE_PATH})
-
-# Add a local CMake module search path dependent on the desired installation destination.
-# Thus, artifacts from the complete source build can be given precendence over any installed versions.
-IF(UNIX)
-    SET (CMAKE_MODULE_PATH "${CMAKE_INSTALL_PREFIX}/share/cmake-${CMAKE_MAJOR_VERSION}.${CMAKE_MINOR_VERSION}/Modules" ${CMAKE_MODULE_PATH})
-ENDIF()
-IF(WIN32)
-    SET (CMAKE_MODULE_PATH "${CMAKE_INSTALL_PREFIX}/CMake-${CMAKE_MAJOR_VERSION}.${CMAKE_MINOR_VERSION}/Modules" ${CMAKE_MODULE_PATH})
-ENDIF()
-
-###########################################################################
-# Include flags for compiling.
-INCLUDE (${CMAKE_CURRENT_SOURCE_DIR}/cmake.Modules/CompileFlags.cmake)
-
-###########################################################################
-# Find OpenDaVINCI.
-SET(OPENDAVINCI_DIR "${CMAKE_INSTALL_PREFIX}")
-FIND_PACKAGE (OpenDaVINCI REQUIRED)
-
-# Find OpenBASN.
-SET(BSN_DIR "${CMAKE_INSTALL_PREFIX}")
-FIND_PACKAGE (OpenBASN REQUIRED)
-
-###########################################################################
-# Set linking libraries to successfully link test suites and binaries.
-SET (LIBRARIES ${OPENDAVINCI_LIBRARIES}
-                ${BSN_LIBRARIES})
-
-###########################################################################
-# Set header files from OpenDaVINCI.
-INCLUDE_DIRECTORIES (${OPENDAVINCI_INCLUDE_DIRS})
-
-# Set header files from OpenBASN.
-INCLUDE_DIRECTORIES (${BSN_INCLUDE_DIRS})
-
-# Set include directory.
-INCLUDE_DIRECTORIES(include)
-
-###########################################################################
-# Build this project.
-FILE(GLOB_RECURSE thisproject-sources "${CMAKE_CURRENT_SOURCE_DIR}/src/*.cpp")
-ADD_LIBRARY (${PROJECT_NAME}lib-static STATIC ${thisproject-sources})
-ADD_EXECUTABLE (${PROJECT_NAME} "${CMAKE_CURRENT_SOURCE_DIR}/apps/${PROJECT_NAME}.cpp")
-TARGET_LINK_LIBRARIES (${PROJECT_NAME} ${PROJECT_NAME}lib-static ${LIBRARIES}) 
-
-###########################################################################
-# Install this project.
-INSTALL(TARGETS ${PROJECT_NAME} RUNTIME DESTINATION bin COMPONENT bodyhub)
