+++ configs/configuration
@@ -1,4 +1,4 @@
-# OpenBASN Global configuration file
+# BSN Global configuration file
 
 ### Global configs
 #global.sensortypes = 3
